Feature/goal

See merge request cloud_track/class_02/web_project3/team05/team05-be!157

+++ src/main/java/com/example/savingsalt/goal/domain/dto/GoalCertificationStatisticsResDto.java
@@ -25,7 +25,7 @@ public class GoalCertificationStatisticsResDto {
     private Set<String> monthlyCertifications;
 
     // 월간 인증 내용 백분율
-    private Map<String, Long> monthlyCertificationPercentages;
+    private Map<String, Double> monthlyCertificationPercentages;
 
     // 일일 인증 금액
     private Long dailyAmount;
